changing memory calculation while blocking hugepages and adding flag validations
